Prepares for npm publish

+++ scripts/version.sh
@@ -1,6 +1,13 @@
 #!/usr/bin/env bash
 
-VERSION=$1
-
-PATH=$(npm bin):$PATH lerna exec -- rm -rf dist
-PATH=$(npm bin):$PATH lerna exec -- npm version ${VERSION}
+if [[ "$1" = "--pre" ]]
+then
+    npm test
+elif [[ "$1" = "--post" ]]
+then
+    git push --tags
+else
+    VERSION=$1
+    PATH=$(npm bin):$PATH lerna exec -- rm -rf dist
+    PATH=$(npm bin):$PATH lerna exec -- npm version ${VERSION}
+fi
